adicionando todas as chamadas a API

+++ app/src/main/java/com/example/weatherapp/data/model/FutureModel.kt
@@ -4,6 +4,6 @@ data class FutureModel(
     val day:String,
     val idPic:Int,
     val status:String,
-    val highTemp:Int,
-    val lowTemp: Int
+    val highTemp:Float,
+    val lowTemp: Float
 )
